Fixing plotly scales.
